0.5.0
Significantly updated the development section of the site to include tabs for sections that open on the same page. Added content for several classes under the development section. Ensured all pages on the site have the favicon for brand consistency. Fixed all the links between files to ensure that they display correctly on the host server.