Version 2.0.0
updated code base for maintainability and readability. also made changes to content and the manner in which the user moves through the site. massively increased the responsiveness of the website and ensured a minimum presentation across all screen sizes and customized presentation on specific popular screen sizes. added functionality to resume page and work page.

note: assets folder was too large to push to github