Version 2.1.0
updated the method used to display portfolio items to the user in order to decrease load time and increase visual presentation. dramatically increased the mobile responsiveness as regards the latest and most popular smartphones and tablets. landscape orientation should now be supported without issue.

note: assets folder to large to push to GitHub